listenerとbot本体をクラス分離
